chore(reactive-forms): move sample to how-to directory
