+++ src/app/data-entries/data-entries-routing.module.ts
@@ -42,7 +42,7 @@ import { InputGroupSample5Component } from './input-group/input-group-sample-5/i
 import { InputGroupSample6Component } from './input-group/input-group-sample-6/input-group-sample-6.component';
 import { InputGroupStyleComponent } from './input-group/input-group-styling/input-group-styling.component';
 import { InputTextSelectionComponent } from './input-group/input-text-selection/input-text-selection.component';
-import { ReactiveFormsSampleComponent } from './input-group/reactive-forms/reactive-forms.component';
+import { ReactiveFormsSampleComponent } from '../how-to/reactive-forms/reactive-forms.component';
 import { RadioGroupVerticalComponent } from './radio/radio-group-vertical/radio-group-vertical.component';
 import { RadioGroupSampleComponent } from './radio/radio-group-sample/radio-group-sample.component';
 import { RadioSample1Component } from './radio/radio-sample-1/radio-sample-1.component';
